[mate][fix] 동행인 상세게시판[작성자 본인일 경우 수정,삭제버튼 o 신청버튼 x][작성자가 아닐경우 신청버튼 o, 조건에 맞을경우 신청메세지, 아닐경우 조건불만족 모달]

+++ src/main/java/com/multi/hontrip/mate/dao/MateDAO.java
@@ -1,10 +1,5 @@
 package com.multi.hontrip.mate.dao;
 
-import com.multi.hontrip.mate.dto.LocationDTO;
-import com.multi.hontrip.mate.dto.MateBoardInsertDTO;
-import com.multi.hontrip.mate.dto.MateBoardListDTO;
-import com.multi.hontrip.mate.dto.PageDTO;
-import org.springframework.stereotype.Repository;
 import com.multi.hontrip.mate.dto.*;
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,9 +34,10 @@ public class MateDAO {
         my.insert("mateBbs.insert", mateBoardInsertDTO);
     }
 
-    public MateBoardInsertDTO mateBoardSelectOne(int id) {
+    public MateBoardInsertDTO mateBoardSelectOne(long id) {
         return my.selectOne("mateBbs.selectOne", id);
     }
+
     public int updateMateBoard(MateBoardInsertDTO mateBoardInsertDTO) {
         return my.update("mateBbs.updateMateBoard", mateBoardInsertDTO);
     }
